refactor: reorganize src/ into discord/, agent/, session/, cli/ subdirectories
